Abstract do_api_call into api.util and default sensor type
