+++ tfc_web/parking/views.py
@@ -1,26 +1,16 @@
-import codecs
 import json
 from datetime import date, timedelta, datetime
-from urllib.request import Request, urlopen
 from urllib.error import HTTPError
-from django.conf import settings
 from django.shortcuts import render
 import logging
+
+from api.util import do_api_call
 
 logger = logging.getLogger(__name__)
 
 #############################################################################
 # Utilities                                                                 £
 #############################################################################
-
-
-def do_api_call(query):
-
-    logger.debug('Query: %s', query)
-    reader = codecs.getreader("utf-8")
-    query = Request(settings.NEW_API_ENDPOINT + query)
-    query.add_header('Authorization', 'Token ' + settings.LOCAL_API_KEY)
-    return json.load(reader(urlopen(query)))
 
 
 def get_parking_list():
